# licne-finansije/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Veb aplikacija za pregled licnih finansija",
        "description": "API za upravljanje ličnim finansijama (autentifikacija, transakcije, kategorije, budžeti, finansijski ciljevi, podsetnici, dokumenta i admin statistike)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registracija korisnika (uz opcioni upload profilne slike) + slanje linka za verifikaciju email-a",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "ime",
                                    "prezime",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "ime": {
                                        "type": "string",
                                        "example": "Marko"
                                    },
                                    "prezime": {
                                        "type": "string",
                                        "example": "Marković"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "marko@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "profilnaSlika": {
                                        "description": "Profilna slika (jpg/jpeg/png, max 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registracija uspesna"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Prijava korisnika (Sanctum token). Ako email nije verifikovan vraca poruku.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "marko@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uspesno prijavljivanje"
                    },
                    "400": {
                        "description": "Email adresa nije verifikovana."
                    },
                    "401": {
                        "description": "Pogresan email ili lozinka"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Odjava korisnika (briše trenutni token)",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Uspesna odjava"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Podaci o trenutno ulogovanom korisniku",
                "operationId": "9d9bc556f927575aa492a88c28fe5908",
                "responses": {
                    "200": {
                        "description": "Podaci o korisniku"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/email/verify/{id}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verifikacija email-a preko potpisanog linka (temporary signed route)",
                "operationId": "23a3f61982dab0cfcd39bf2e009ced8a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID korisnika",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "expires",
                        "in": "query",
                        "description": "Timestamp isteka linka (automatski u signed URL-u)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "signature",
                        "in": "query",
                        "description": "Potpis linka (automatski u signed URL-u)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email verifikovan (ili vec verifikovan)"
                    },
                    "401": {
                        "description": "Link nevazeći ili je istekao"
                    }
                }
            }
        },
        "/api/update-profile": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Azuriranje profila korisnika (ime, prezime, email, password, profilna slika)",
                "operationId": "630da0bb83fd6b02c963532ebc1fbc19",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "ime": {
                                        "type": "string",
                                        "example": "Marko"
                                    },
                                    "prezime": {
                                        "type": "string",
                                        "example": "Marković"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "marko_novo@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "novasifra123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "novasifra123"
                                    },
                                    "profilnaSlika": {
                                        "description": "Nova profilna slika (jpg/jpeg/png, max 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil uspesno azuriran"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}